Safari-609.1.11.1/LayoutTests/accessibility/ios-simulator/menuitem-checked-expected.txt


Item one
Item two
Item three
Item four
Item five
This tests that menu items can report the right value whether it's checked.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS radio.traits != menuItem1.traits is true
PASS menuItem1.stringValue is 'AXValue: 1'
PASS menuItem2.stringValue is 'AXValue: 0'
PASS menuItem1.traits == menuItem3.traits is true
PASS menuItem3.stringValue is 'AXValue: 1'
PASS menuItem4.stringValue is 'AXValue: 0'
PASS menuItem5.traits != menuItem4.traits is true
PASS successfullyParsed is true

TEST COMPLETE
